Ensure data type of urllib.error.HTTPError response to be string
