+++ codalab/client/json_api_client.py
@@ -4,7 +4,7 @@ import sys
 import six
 import urllib.request, urllib.parse, urllib.error
 
-from codalab.common import http_error_to_exception, precondition, UsageError
+from codalab.common import http_error_to_exception, precondition, ensure_str, UsageError
 from codalab.worker.rest_client import RestClient, RestClientException
 from codalab.worker.download_util import BundleTarget
 
@@ -16,7 +16,7 @@ def wrap_exception(message):
                 return f(*args, **kwargs)
             except urllib.error.HTTPError as e:
                 # Translate known errors to the standard CodaLab errors
-                error_body = e.read()
+                error_body = ensure_str(e.read())
                 exc = http_error_to_exception(e.code, error_body)
                 # All standard CodaLab errors are subclasses of UsageError
                 if isinstance(exc, UsageError):
